Add wheelie package
The wheelie package is a library wrapper around helm, which
provides three primary methods: `EnsureReleasePresent`,
`EnsureReleaseAbsent`, and `EnsureReleasePurged`.

The purpose is to provide idempotent versions of the operations
that helm exposes. For example, `EnsureReleasePresent` will call
helm's `Install` method if a release is not present, but `Upgrade`
if it is present. It will also do nothing if there are no changes
to make,  rather than creating a new release version as helm
upgrade does.